requirements.txt
streamlit
pillow==9.2.0
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
